requirements.txt
blessings==1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
drawille==0.1.0
gpustat==0.6.0
h5py==2.10.0
hdf5storage==0.1.15
huepy==1.2.1
idna==2.8
npyscreen==4.10.5
numpy==1.17.3
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==5.3.0
protobuf==3.11.2
psutil==5.6.7
ptop==1.0
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
scipy==1.2.1
six==1.13.0
tensorboardX==1.9
torch==1.3.1
torchvision==0.4.2
urllib3==1.25.7
yacs==0.1.6
